+++ docs/changelog.md
@@ -3,7 +3,9 @@
 ### Features
 - The "Starlark code successfully executed" or "Error encountered running Starlark code" messages are now "Starlark 
 code successfully run in dry-run mode" and "Error encountered running Starlark code in dry-run mode" when Starlark is 
-run in dry-run mode (and without the "in dry-run mode" when the script is executed for real)
+run in dry-run mode (and without the "in dry-run mode" when the script is executed for real)\
+- Added `RunStarlarkScriptBlocking`, `RunStarlarkPackageBlocking` and `RunStarlarkRemotePackageBlocking` functions
+to the enclave context to facilitate automated testing in our current modules.
 
 ### Fixes
 - Don't duplicate instruction position information in `store_service_files`